+++ Assets/LDtkUnity/Samples~/Common/Scripts/ExampleMove.cs
@@ -1,5 +1,9 @@
 ﻿using UnityEngine;
 
+#if ENABLE_INPUT_SYSTEM
+using UnityEngine.InputSystem;
+#endif
+
 namespace Samples
 {
     public class ExampleMove : MonoBehaviour
@@ -20,9 +24,39 @@ namespace Samples
 
         private Vector2 GetMove()
         {
-            float x = Input.GetAxisRaw("Horizontal");
-            float y = Input.GetAxisRaw("Vertical");
-            return new Vector2(x, y).normalized;
+            Vector2 move = Vector2.zero;
+            
+#if ENABLE_INPUT_SYSTEM
+            Keyboard keyboard = Keyboard.current;
+            Gamepad gamepad = Gamepad.current;
+            if (keyboard != null && keyboard.anyKey.isPressed)
+            {
+                if (keyboard.wKey.isPressed || keyboard.upArrowKey.isPressed)
+                    move.y += 1;
+                
+                if (keyboard.aKey.isPressed || keyboard.leftArrowKey.isPressed)
+                    move.x -= 1;
+                
+                if (keyboard.sKey.isPressed || keyboard.downArrowKey.isPressed)
+                    move.y -= 1;
+                
+                if (keyboard.dKey.isPressed || keyboard.rightArrowKey.isPressed)
+                    move.x += 1;
+            }
+            else if (gamepad != null)
+            {
+                if (gamepad.leftStick.IsActuated())
+                    move = gamepad.leftStick.ReadValue();
+                
+                else if (gamepad.dpad.IsActuated())
+                    move = gamepad.dpad.ReadValue();
+            }
+#elif ENABLE_LEGACY_INPUT_MANAGER
+            move.x = Input.GetAxisRaw("Horizontal");
+            move.y = Input.GetAxisRaw("Vertical");
+#endif
+            
+            return move.normalized;
         }
     }
 }