+++ Assets/LDtkUnity/CHANGELOG.md
@@ -14,6 +14,7 @@ This update focuses on significant performance optimisation and a revamp to sepa
   
 - Added safe type checking and support for external enums
 - All debug logs will now only show a maximum of 50 duplicates, and has colored text in the console
+- Sample scripts now support the new Input System
 
 ## Changes
 - Changed how the project importer manages assets